Added backup system and readme

+++ start.py
@@ -2,6 +2,7 @@ import atexit
 import io
 import os.path
 import subprocess
+import time
 from dotenv import load_dotenv
 import os
 import json
@@ -14,34 +15,77 @@ from google_auth_oauthlib.flow import InstalledAppFlow
 from googleapiclient.discovery import build
 from googleapiclient.errors import HttpError
 from googleapiclient.http import MediaIoBaseDownload
+from googleapiclient.http import MediaFileUpload
 
 # If modifying these scopes, delete the file token.json.
 SCOPES = ["https://www.googleapis.com/auth/drive"]
 
-def run_mc_server_as_subprocess():
-    start_cmd = "java -Xmx6G -jar fabric-server-launch.jar nogui"
-    return subprocess.Popen(start_cmd, shell=True)
+def getLatestLocalBackup(LOCAL_SERVER_DIR):
+
+    if not os.path.exists(LOCAL_SERVER_DIR):
+        return -1
+
+    # Get all files in the directory
+    files = os.listdir(LOCAL_SERVER_DIR)
+    # Filter out all files that are not .zip
+    files = [f for f in files if f.endswith(".zip")]
+    # Sort the files by name
+    files.sort()
+    # Return the last file
+    return getBackupIterationFromName(files[-1])
+
+def getBackupIterationFromName(name):
+    # Try to get number from name, else return -1
+    try:
+        return int(name.split(".")[0])
+    except:
+        return -1
+
+def run_mc_server_as_subprocess(LOCAL_SERVER_DIR):
+    start_cmd = ["java", "-Xmx6G", "-jar", "fabric-server-launch.jar", "nogui"]
+    global server_process_global
+    server_process_global = subprocess.Popen(start_cmd, cwd=LOCAL_SERVER_DIR, stdin=subprocess.PIPE)
+    print("Minecraft server started.")
+
+    # Print the pids
+    print(f"Server PID: {server_process_global.pid}")
+    print(f"Server Process Group ID: {os.getpgid(server_process_global.pid)}")
+    
+    #Print this script's PID
+    print(f"Script PID: {os.getpid()}")
+    print(f"Script Process Group ID: {os.getpgrp()}")
 
 def stop_server():
-    global server_process
-    if server_process:
+    if server_process_global:
         print("Stopping minecraft server...")
-        server_process.stdin.write('/stop\n')
-        server_process.stdin.flush()
-        server_process.wait()  # Wait for the server to finish
+        server_process_global.stdin.write('/stop\n'.encode())
+        server_process_global.stdin.flush()
+        server_process_global.wait()  # Wait for the server to finish
         print("Minecraft server stopped.")
+    else:
+        print("ERROR: Server process could not be found.")
+
+def clear_zips(directory):
+    for filename in os.listdir(directory):
+        if filename.endswith(".zip"):
+            file_path = os.path.join(directory, filename)
+            try:
+                os.unlink(file_path)
+            except Exception as e:
+                print(f"Failed to delete {file_path}. Reason: {e}")
 
 def clear_directory(directory):
+    if not os.path.exists(directory):
+        return
     for filename in os.listdir(directory):
         file_path = os.path.join(directory, filename)
         try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
+            if os.path.isfile(file_path) or os.path.islink(file_path) and not file_path.endswith(".zip"):
                 os.unlink(file_path)
             elif os.path.isdir(file_path):
                 shutil.rmtree(file_path)
         except Exception as e:
             print(f"Failed to delete {file_path}. Reason: {e}")
-
 
 def download_file(
     service, file_id, destination_folder, destination_file_name="server.zip"
@@ -89,34 +133,13 @@ def build_directory_structure(service, folder_id, indent=""):
             metadata_results[item["name"]] = build_directory_structure(
                 service, item["id"], indent + "  "
             )
+            metadata_results[item["name"]]["id"] = item["id"]
         else:
             metadata_results[item["name"]] = item
 
     return metadata_results
 
-
-def print_directory_structure(service, folder_id, indent=""):
-    # List all files and folders in the current folder
-    results = (
-        service.files()
-        .list(
-            q=f"'{folder_id}' in parents",
-            fields="nextPageToken, files(id, name, mimeType)",
-        )
-        .execute()
-    )
-    items = results.get("files", [])
-
-    for item in items:
-        print(f"{indent}{item['name']}")
-
-        # If the item is a folder, recursively print its contents
-        if item["mimeType"] == "application/vnd.google-apps.folder":
-            print_directory_structure(service, item["id"], indent + "  ")
-
-
 def main():
-
     load_dotenv()
 
     """Shows basic usage of the Drive v3 API.
@@ -169,6 +192,7 @@ def main():
         # Each backup is 0.zip, 1.zip, 2.zip, etc.
         # Get all keys and sort them
         latest_backups_keys = list(latest_backups.keys())
+        latest_backups_keys.remove("id")
         latest_backups_keys.sort()
         latest_backup = latest_backups_keys[-1]
 
@@ -177,9 +201,8 @@ def main():
         latest_backup_name = latest_backups[latest_backup]["name"]
 
         # If the latest backup is already downloaded, skip this step
-        if os.path.exists(f"{LOCAL_SERVER_DIR}/{latest_backup_name}"):
+        if getLatestLocalBackup(LOCAL_SERVER_DIR) >= getBackupIterationFromName(latest_backup_name):
             print(f"Latest backup already downloaded: {latest_backup_name}")
-            return
         else:
             clear_directory(LOCAL_SERVER_DIR)
             download_file(
@@ -193,10 +216,42 @@ def main():
                 zip_ref.extractall(LOCAL_SERVER_DIR)
 
         # Start the server
-        server_process = run_mc_server_as_subprocess()
+        run_mc_server_as_subprocess(LOCAL_SERVER_DIR)
         atexit.register(stop_server)
 
-        # Schedule backups
+        # Schedule backups while the server is running
+        while server_process_global and server_process_global.poll() is None:
+            time.sleep(int(BACKUP_INTERVAL))
+
+            # Stop the server
+            stop_server()
+
+            # Get index of latest backup from name
+            latest_backup_index = getLatestLocalBackup(LOCAL_SERVER_DIR)
+            next_index = latest_backup_index + 1
+            
+            print(f"Creating backup #{next_index}")
+
+            # Remove current zip
+            clear_zips(LOCAL_SERVER_DIR)
+
+            # Zip the contents of the server directory
+            shutil.make_archive(f"{LOCAL_SERVER_DIR}/{next_index}", "zip", LOCAL_SERVER_DIR)
+
+            # Upload the zip file to the backups folder
+            file_metadata = {
+                "name": f"{next_index}.zip",
+                "parents": [metadata["backups"][SERVER_NAME]["id"]],
+            }
+            media = MediaFileUpload(
+                f"{LOCAL_SERVER_DIR}/{next_index}.zip", mimetype="application/zip"
+            )
+            file = service.files().create(body=file_metadata, media_body=media, fields="id").execute()
+            print(f"Created backup file: {next_index}.zip %s" % file.get("id"))
+            latest_backup_name = f"{next_index}.zip"
+
+            # Start the server
+            run_mc_server_as_subprocess(LOCAL_SERVER_DIR)
 
     except HttpError as error:
         # TODO(developer) - Handle errors from drive API.
